backend/requirements.txt
fastapi[standard]
sqlalchemy
passlib
python-jose[cryptography]
bcrypt==4.0.1